refactor(open-file): move mac.rs tests into sibling mac_test.rs
Match the directory's existing convention (mod.rs already wires
mod_test.rs the same way) and the dominant repo-wide pattern of
using `#[path = "..._test.rs"] mod tests;` rather than inline
`mod tests` blocks.

No behavior change.
